SmartHomeAssistant/requirements.txt
asttokens==2.4.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-aiplatform==1.41.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
jedi==0.19.1
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
SpeechRecognition==3.10.1
stack-data==0.6.3
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
vertexai==0.0.1
wcwidth==0.2.13
